chore: Convert CLI throw test to Jest / TS
This patch merges the functional CLI init and CLI exec tests into a
single CLI functional test, and makes the init test Jest / TS
